requirements.txt
annotated-types==0.7.0
anyio==4.6.2
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.1
debugpy==1.8.7
distro==1.9.0
Django==5.1.2
django-cors-headers==4.5.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jwcrypto==1.5.6
oauthlib==3.2.2
openai==1.51.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3